backend/fastapi_env/requirements.txt
aiofiles==24.1.0
aiohttp==3.10.10
anthropic==0.39.0
fastapi==0.115.6
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-cloud-translate==3.16.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
groq==0.15.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx
httpx-sse==0.4.0
imagesize==1.4.1
langchain==0.3.7
langchain-anthropic==0.2.4
langchain-community==0.3.5
langchain-core==0.3.31
langchain-google-genai==2.0.4
langchain-groq==0.2.3
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langdetect==1.0.9
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-checkpoint-sqlite==2.0.1
langgraph-sdk==0.1.36
numpy==1.26.4
openai==1.53.0
pandas==2.2.2
pdf2image==1.17.0
pdfservices-sdk==4.1.0
pillow==10.3.0
pinecone==6.0.1
py7zr==0.22.0
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.6.1
PyMuPDF==1.25.1
pymupdf4llm==0.0.17
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.9.11
requests==2.31.0
rsa==4.9
sse-starlette==2.2.1
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
urllib3==2.2.1
uvicorn==0.32.0
python-multipart==0.0.20
pyarrow==19.0.0